% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_timeseries.R
\name{abun_data_bycell}
\alias{abun_data_bycell}
\title{Function to extract cell-specific abundance data}
\usage{
abun_data_bycell(abun_data, n_small_min = 10)
}
\arguments{
\item{abun_data}{abundance data}

\item{n_small_min}{minimum number of small cells to compute abundance index for large cell}
}
\value{
data.frame of abundances
}
\description{
Function to extract cell-specific abundance data
}
