% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daymet_extract.R
\name{daymet_extract}
\alias{daymet_extract}
\title{extract Daymet summaries over a spatiotemporal cell of interest}
\usage{
daymet_extract(
  cell = 621,
  res = 6,
  variable = "swe",
  mindate = "2014-01-01",
  maxdate = "2014-03-01"
)
}
\arguments{
\item{cell}{the cell id from a dggrid}

\item{res}{the resolution of the dggrid}

\item{variable}{the name of the daymet variable}

\item{mindate}{the minimum date over which to summarize}

\item{maxdate}{the maximum date over which to summarize}
}
\value{
the median (over pixels in the spatial cell) of the mean (over days
within the pixels) of the variable of interest.
}
\description{
extract Daymet summaries over a spatiotemporal cell of interest
}
