% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_and_bootstrap.R
\name{get_abun}
\alias{get_abun}
\title{Get abundance data}
\usage{
get_abun(sp_data, n_rep, .cores = 4)
}
\arguments{
\item{sp_data}{Data for a particular species and year from get_grid_data()}
}
\value{
a dataframe
}
\description{
Get abundance data
}
