% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBS_functions.R
\name{get_abun_bcr}
\alias{get_abun_bcr}
\title{get bootstrapped abundance estimates for bcrs}
\usage{
get_abun_bcr(sp_data, bcrs, n_rep, .cores = 4)
}
\arguments{
\item{sp_data}{species abundance data outputted by get_bcr_data}

\item{bcrs}{sf object holding the bcrs}

\item{n_rep}{number of bootstrap replicates}

\item{.cores}{number of cores for parallel processing}
}
\value{
matrix of bootstrapped abundances
}
\description{
get bootstrapped abundance estimates for bcrs
}
