% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_timeseries.R
\name{get_abun_summary}
\alias{get_abun_summary}
\title{Abundance summary across years}
\usage{
get_abun_summary(abun_data, n_small_min)
}
\arguments{
\item{abun_data}{abundance data}

\item{n_small_min}{minimum number of small cells to compute abundance index for large cell}
}
\value{
something
}
\description{
Abundance summary across years
}
