% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grid_data.R
\name{get_cell_data}
\alias{get_cell_data}
\title{Get data for a stixel}
\usage{
get_cell_data(zfd, t, all_cells_small)
}
\arguments{
\item{zfd}{zero-filled data}

\item{t}{the tgrid element}

\item{all_cells_small}{the list of all cells}
}
\value{
row vector
}
\description{
Get data for a stixel
}
