% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grid_data.R
\name{get_grid_data}
\alias{get_grid_data}
\title{Get eBird summaries on cells of a small grid, tagged with membership in a large grid.}
\usage{
get_grid_data(
  data,
  .year,
  tgrid_min,
  tgrid_max,
  time_window = "gridded",
  min_lat,
  max_lat,
  min_lon,
  max_lon = Inf,
  large_grid = 6,
  small_grid = 11,
  time_grid = 7,
  .cores = 4
)
}
\arguments{
\item{data}{zero-filled data (output from import_from_erd())}

\item{.year}{the year}

\item{tgrid_min}{the minimum time grid index (week number if \code{time_grid=7}).}

\item{tgrid_max}{the maximum time grid index (week number if \code{time_grid=7}).}

\item{time_window}{"gridded" or "full"}

\item{min_lat}{the minimum latitude (decimal degrees) where the prediction is desired}

\item{max_lat}{the maximum latitude}

\item{min_lon}{the minimum longitude}

\item{max_lon}{the maximum longitude}

\item{large_grid}{the dggridR resolution of the large grid}

\item{small_grid}{the dggridR resolution of the small grid}

\item{time_grid}{the temporal grid size in days (default 7, one week).}

\item{.cores}{number of parallel cores for \code{get_cell_data()} calls}
}
\description{
Get eBird summaries on cells of a small grid, tagged with membership in a large grid.
}
