% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BBS_functions.R
\name{get_pixels_bcr}
\alias{get_pixels_bcr}
\title{Get the small pixels corresponding to each BCR, each of which represents a large pixel}
\usage{
get_pixels_bcr(bcrs, dg_small)
}
\arguments{
\item{bcrs}{sf representation of BCRs}

\item{dg_small}{a dggrid representing the small pixels}
}
\value{
dataframe giving the bcr for each small cell
}
\description{
Get the small pixels corresponding to each BCR, each of which represents a large pixel
}
