% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median_passage.R
\name{median_passage}
\alias{median_passage}
\title{Compute median passage dates for a species}
\usage{
median_passage(sp_data, season, min_day, max_day, lat1, lat2, bandwidth = 1)
}
\arguments{
\item{sp_data}{species data returned from import_from_erd}

\item{lat1}{southern boundary of ROI}

\item{lat2}{northern boundary of ROI}

\item{bandwidth}{width of latitudinal bands to get median, in degrees}
}
\value{
dataframe
}
\description{
Compute median passage dates for a species
}
