% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_codes.R
\docType{data}
\name{species_codes}
\alias{species_codes}
\title{four and six letter bird species codes for focal species}
\format{
A data.frame with 22 rows and 2 columns
}
\source{
this script
}
\usage{
species_codes
}
\description{
four and six letter bird species codes for focal species
}
\keyword{datasets}
