% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grid_data.R
\name{stixel_mean_small}
\alias{stixel_mean_small}
\title{Get mean value for a stixel}
\usage{
stixel_mean_small(x)
}
\arguments{
\item{x}{data for a single cell}
}
\value{
average abundance for the cell
}
\description{
Get mean value for a stixel
}
