% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_results.R
\name{summarize_avg_and_rep}
\alias{summarize_avg_and_rep}
\title{Summarize computed results from get_abun}
\usage{
summarize_avg_and_rep(
  spring,
  fall,
  year_remove = 4,
  remove = T,
  min_small = 1,
  min_list = 1,
  ci = 0.8
)
}
\arguments{
\item{spring}{output from get_abun for a spring}

\item{fall}{output from get_abun for a fall of the same year}

\item{year_remove}{remove all stixels that are NA in any year after
year_remove.  Remove these stixels from ALL years}

\item{remove}{if T, remove all years after year_remove.  Useful scatterplots
of north versus south or index versus duration, where
the year information isn't visually apparent}

\item{min_small}{the minimum number of small stixels with at least one checklist
require to include a big stixel in the analysis}

\item{min_list}{the minimum number of checklists required to include a big
stixel in the analysis.  Guaranteed to have no effect if
less than or equal to min_small}

\item{ci}{Credible interval width for quantities reported with credible intervals.}
}
\value{
a list of means and CIs for spring and fall indices and ratios
}
\description{
Summarize computed results from get_abun
}
