% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_timeseries.R
\name{use_cell_years}
\alias{use_cell_years}
\title{function to determine which cell-years are analyzeable}
\usage{
use_cell_years(
  ratio_series,
  uncertainty_high_grade = Inf,
  inf_exclude = F,
  element_1_exclude = T,
  n_min_prod = 5,
  n_min_surv = 5,
  n_min_full = 5
)
}
\arguments{
\item{ratio_series}{a ratio series}

\item{uncertainty_high_grade}{maximum allowable uncertainty in cell-year}

\item{inf_exclude}{exclude all cell years when the ratio is infinite in any year}

\item{element_1_exclude}{exclude the first year of the series}

\item{n_min_prod}{minimimum number of productivity indices necessary to include
any years/seasons at all}

\item{n_min_surv}{minimum number of survicial indices necessary to include any
years/seasons at all}

\item{n_min_full}{minimum number of individual years with both a productivity and
a survical index to include any years/seasons at all}
}
\value{
a logical vector of which elements to include
}
\description{
function to determine which cell-years are analyzeable
}
