% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grid_data.R
\name{get_grid_data}
\alias{get_grid_data}
\title{Get eBird summaries on cells of a small grid, tagged with membership in a large grid.}
\usage{
get_grid_data(
  data,
  .year,
  tgrid_min,
  tgrid_max,
  min_lat,
  max_lat,
  min_lon,
  large_grid = 6,
  small_grid = 11,
  .cores = 4
)
}
\arguments{
\item{data}{zero-filled data (output from import_from_erd())}

\item{.year}{the year}

\item{tgrid_min}{the minimum week of year}

\item{tgrid_max}{the maximum week of year}

\item{min_lat}{the minimum latitude (decimal degrees) where the prediction is desired}

\item{max_lat}{the maximum latitude}

\item{min_lon}{the minimum longitude}

\item{large_grid}{the dggridR resolution of the large grid}

\item{small_grid}{the dggridR resolution of the small grid}
}
\description{
Get eBird summaries on cells of a small grid, tagged with membership in a large grid.
}
