% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ratios.R, R/sample_grid_abun.R
\name{get_ratios}
\alias{get_ratios}
\title{Get log-ratios}
\usage{
get_ratios(data, cells_all, period = c("spring", "fall"))

get_ratios(data, cells_all, period = c("spring", "fall"))
}
\arguments{
\item{data}{abundance index data}

\item{cells_all}{cells to compute}

\item{period}{period to compute}
}
\description{
Get log-ratios

Get log-ratios
}
